Remove used switch entries
Removing all the switch entries cause the Dashboard was referencing
switch.family_media_center and switch.family_room_tv but this was somehow
connected with Jojo's Train Light.

Since we are not using the automation on switches right now, disable it
to prevent it acting on switches we don't expect.